updated comment
The number of trees built per gardener doesn't significantly affect the outcome.
(experimented with n=7, n=8, n=9 with benchmark script)

I'm guessing this is partly because of cluttering late in the game--beyond a
certain point, it's random whether a gardener will be able to add trees or not.
Maybe the result would be different if we spaced gardeners differently.