significantly reworked micro, pathing, aggresion,
and build order. This just tries to make a really huge army and disregards
most of the VP stuff. Realistically, once we have a big army, we ought to be
able to save VP and go for a VP win. But it's hard to say how big an army we
need.